+++ R/allClasses.R
@@ -8,21 +8,25 @@ setClass(
         normY = "character"
     )
 )
+setClassUnion("numericOrNULL", c("character", "NULL"))
 setClass(
     "SbivarResultsMoransI",
     contains = "SbivarResults",
     slots = c(
-        maxIxy = "numeric",
+        maxIxy = "numericOrNULL",
         estimateSEsMoransI = "logical",
         wo = "character",
         wParams = "numeric"
     )
 )
+setOldClass("corSpatial")
+setClassUnion("corSpatialOrNULL", c("character", "NULL"))
 setClass(
-    "SbivarResultsGAM",
+    "SbivarResultsGAMs",
     contains = "SbivarResults",
     slots = c(
-        families = "character",
-        correlation = "logical"
+        families = "list",
+        correlation = "corSpatialOrNULL",
+        Gamm = "logical"
     )
 )